requirements.txt
numpy~=1.24.3
tqdm~=4.60.0
pandas~=1.2.3
seaborn~=0.12.2
matplotlib>=3.3.4
pygame~=2.0.1
networkx>=2.6.3
simplejson~=3.17.2
PyYAML~=5.3.1
natsort~=7.1.1
torch~=1.10.0
gymnasium~=0.28.1